svn path=/branches/ros-amd64-bringup/; revision=36862

+++ reactos/include/ddk/winddk.h
@@ -7763,7 +7763,7 @@ VOID
 NTAPI
 ProbeForRead(
   IN CONST VOID  *Address,
-  IN ULONG  Length,
+  IN SIZE_T  Length,
   IN ULONG  Alignment);
 
 NTKERNELAPI
@@ -7771,7 +7771,7 @@ VOID
 NTAPI
 ProbeForWrite(
   IN PVOID  Address,
-  IN ULONG  Length,
+  IN SIZE_T  Length,
   IN ULONG  Alignment);
